fix/UT01 - UserFilterRepositoryTest 브랜치 커버리지 만족하도록 수정 (#553)
* refactor: Json Pretty Print 설정을 local, test에서만 적용하도록 변경

* feat: query pattern 변경

* refactor: file과 console 패턴 분리

* feat: AsyncAppender 도입

* refactor: utf8 제거

* refactor: AsyncAppender 설정 변경

* refactor: Config 파일 분리

* fix: branch coverage 만족하도록 테스트 추가
