+++ backend/src/test/java/botobo/core/application/WorkbookServiceTest.java
@@ -171,11 +171,11 @@ class WorkbookServiceTest {
                 .id(1L)
                 .name("피케이의 공유 문제집")
                 .cards(new Cards(List.of(
-                        Card.builder()
-                                .id(1L)
-                                .question("question")
-                                .answer("answer")
-                                .build())
+                                Card.builder()
+                                        .id(1L)
+                                        .question("question")
+                                        .answer("answer")
+                                        .build())
                         )
                 )
                 .opened(true)
@@ -205,11 +205,11 @@ class WorkbookServiceTest {
                 .id(1L)
                 .name("피케이의 공유 문제집")
                 .cards(new Cards(List.of(
-                        Card.builder()
-                                .id(1L)
-                                .question("question")
-                                .answer("answer")
-                                .build())
+                                Card.builder()
+                                        .id(1L)
+                                        .question("question")
+                                        .answer("answer")
+                                        .build())
                         )
                 )
                 .opened(true)
@@ -239,11 +239,11 @@ class WorkbookServiceTest {
                 .id(1L)
                 .name("피케이의 공유 문제집")
                 .cards(new Cards(List.of(
-                        Card.builder()
-                                .id(1L)
-                                .question("question")
-                                .answer("answer")
-                                .build())
+                                Card.builder()
+                                        .id(1L)
+                                        .question("question")
+                                        .answer("answer")
+                                        .build())
                         )
                 )
                 .opened(false)
